Move symbols into a separate file
